[refactor] Move literal construction to expr module
